moduletest: Remove the terraform.io/builtin/test provider
This was a temporary placeholder for some first-class syntax for
describing test assertions. We're now going to pivot to using the various
built-in "check" features: smoke tests, preconditions, and postconditions.

As of this commit "terraform test" doesn't yet know how to treat check
failures as test failures, and so this will be effective only for
preconditions/postconditions and they'll cause errors rather than real
test failures. In subsequent commits we'll replace the temporary modeling
of test results with an equivalent using the same checks state objects
we use in other spots, so that we can report test failures in the same
way that we'd report runtime check failures.